Modified the device orientation profile
