+++ dConnectDevicePlugin/dConnectHost/src/org/deviceconnect/android/deviceplugin/host/HostDeviceService.java
@@ -44,8 +44,6 @@ import org.deviceconnect.android.event.cache.MemoryCacheController;
 import org.deviceconnect.android.localoauth.LocalOAuth2Main;
 import org.deviceconnect.android.message.DConnectMessageService;
 import org.deviceconnect.android.message.MessageUtils;
-import org.deviceconnect.android.profile.DConnectProfile;
-import org.deviceconnect.android.profile.DeviceOrientationProfile;
 import org.deviceconnect.android.profile.FileDescriptorProfile;
 import org.deviceconnect.android.profile.MediaPlayerProfile;
 import org.deviceconnect.android.profile.ServiceDiscoveryProfile;
@@ -65,10 +63,6 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
 import android.database.Cursor;
-import android.hardware.Sensor;
-import android.hardware.SensorEvent;
-import android.hardware.SensorEventListener;
-import android.hardware.SensorManager;
 import android.media.AudioManager;
 import android.media.MediaPlayer;
 import android.media.MediaPlayer.OnCompletionListener;
@@ -87,8 +81,7 @@ import android.webkit.MimeTypeMap;
  * 
  * @author NTT DOCOMO, INC.
  */
-public class HostDeviceService extends DConnectMessageService implements
-        SensorEventListener {
+public class HostDeviceService extends DConnectMessageService {
     /** マルチキャスト用のタグ. */
     private static final String HOST_MULTICAST = "deviceplugin.host";
 
@@ -100,9 +93,6 @@ public class HostDeviceService extends DConnectMessageService implements
 
     /** バッファーサイズ. */
     private static final int BUFFER_SIZE = 1024;
-
-    /** SensorManager. */
-    private SensorManager mSensorManager;
 
     /** ServiceID. */
     private String mServiceId;
@@ -404,212 +394,7 @@ public class HostDeviceService extends DConnectMessageService implements
     //
     // Device Orientation Profile
     //
-    /** 加速度 x. */
-    private float mAccellX;
 
-    /** 加速度 y. */
-    private float mAccellY;
-
-    /** 加速度 z. */
-    private float mAccellZ;
-
-    /** Gyro x. */
-    private float mGyroX;
-
-    /** Gyro y. */
-    private float mGyroY;
-
-    /** Gyro z. */
-    private float mGyroZ;
-
-    /** 前回の加速度の計測時間を保持する. */
-    private long mAccelStartTime;
-
-    /**
-     * Device Orientationのキャッシュを残す時間を定義する.
-     */
-    private static final long DEVICE_ORIENTATION_CACHE_TIME = 100;
-
-    /**
-     * Device Orientationのデータを取得する.
-     * @param response データを格納するレスポンス
-     * @return trueの場合には即座に値を返却する、falseの場合には返さない
-     */
-    public boolean getDeviceOrientationEvent(final Intent response) {
-        long t = System.currentTimeMillis() - mAccelStartTime;
-        if (t > DEVICE_ORIENTATION_CACHE_TIME) {
-            List<Sensor> sensors;
-            final SensorEventListener l = new SensorEventListener() {
-                @Override
-                public void onSensorChanged(final SensorEvent event) {
-                    if (event.sensor.getType() == Sensor.TYPE_ACCELEROMETER) {
-                        mAccellX = event.values[0];
-                        mAccellY = event.values[1];
-                        mAccellZ = event.values[2];
-
-                        Bundle orientation = createOrientation();
-                        DConnectProfile.setResult(response, DConnectMessage.RESULT_OK);
-                        DeviceOrientationProfile.setOrientation(response, orientation);
-                        sendResponse(response);
-
-                        mSensorManager.unregisterListener(this);
-                    } else if (event.sensor.getType() == Sensor.TYPE_GYROSCOPE) {
-                        mGyroX = event.values[0];
-                        mGyroY = event.values[1];
-                        mGyroZ = event.values[2];
-                    }
-                }
-                @Override
-                public void onAccuracyChanged(final Sensor sensor, final int accuracy) {
-                }
-            }; 
-
-            mSensorManager = (SensorManager) getSystemService(SENSOR_SERVICE);
-            sensors = mSensorManager.getSensorList(Sensor.TYPE_ACCELEROMETER);
-            if (sensors.size() > 0) {
-                Sensor sensor = sensors.get(0);
-                mSensorManager.registerListener(l, sensor,
-                        SensorManager.SENSOR_DELAY_NORMAL);
-                mAccelStartTime = System.currentTimeMillis();
-            } else {
-                MessageUtils.setNotSupportAttributeError(response);
-                return true;
-            }
-
-            sensors = mSensorManager.getSensorList(Sensor.TYPE_GYROSCOPE);
-            if (sensors.size() > 0) {
-                Sensor sensor = sensors.get(0);
-                mSensorManager.registerListener(l, sensor,
-                        SensorManager.SENSOR_DELAY_NORMAL);
-            } else {
-                MessageUtils.setNotSupportAttributeError(response);
-                return true;
-            }
-
-            return false;
-        } else {
-            Bundle orientation = createOrientation();
-            DConnectProfile.setResult(response, DConnectMessage.RESULT_OK);
-            DeviceOrientationProfile.setOrientation(response, orientation);
-            return true;
-        }
-    }
-
-    /**
-     * Device Orientation Profile<br>
-     * イベントの登録.
-     * 
-     * @param response
-     *            レスポンス
-     * @param serviceId
-     *            サービスID
-     * @param sessionKey
-     *            セッションキー
-     */
-    public void registerDeviceOrientationEvent(final Intent response,
-            final String serviceId, final String sessionKey) {
-
-        mServiceId = serviceId;
-        mSensorManager = (SensorManager) getSystemService(SENSOR_SERVICE);
-
-        List<Sensor> sensors = mSensorManager
-                .getSensorList(Sensor.TYPE_ACCELEROMETER);
-        if (sensors.size() > 0) {
-            Sensor sensor = sensors.get(0);
-            mSensorManager.registerListener(this, sensor,
-                    SensorManager.SENSOR_DELAY_NORMAL);
-            mAccelStartTime = System.currentTimeMillis();
-        } else {
-            MessageUtils.setNotSupportAttributeError(response);
-            return;
-        }
-
-        sensors = mSensorManager.getSensorList(Sensor.TYPE_GYROSCOPE);
-        if (sensors.size() > 0) {
-            Sensor sensor = sensors.get(0);
-            mSensorManager.registerListener(this, sensor,
-                    SensorManager.SENSOR_DELAY_NORMAL);
-        } else {
-            MessageUtils.setNotSupportAttributeError(response);
-            return;
-        }
-
-        DConnectProfile.setResult(response, DConnectMessage.RESULT_OK);
-        response.putExtra(DConnectMessage.EXTRA_VALUE,
-                "Register OnDeviceOrientation event");
-    }
-
-    /**
-     * Device Orientation Profile イベントの解除.
-     * @param response レスポンス
-     */
-    public void unregisterDeviceOrientationEvent(final Intent response) {
-        mSensorManager.unregisterListener(this);
-        response.putExtra(DConnectMessage.EXTRA_RESULT,
-                DConnectMessage.RESULT_OK);
-        response.putExtra(DConnectMessage.EXTRA_VALUE,
-                "Unregister OnDeviceOrientation event");
-    }
-
-    /**
-     * Orientationのデータを作成する.
-     * @return Orientationのデータ
-     */
-    private Bundle createOrientation() {
-        long interval = System.currentTimeMillis() - mAccelStartTime;
-
-        Bundle orientation = new Bundle();
-        Bundle a1 = new Bundle();
-        DeviceOrientationProfile.setX(a1, 0.0);
-        DeviceOrientationProfile.setY(a1, 0.0);
-        DeviceOrientationProfile.setZ(a1, 0.0);
-
-        Bundle a2 = new Bundle();
-        DeviceOrientationProfile.setX(a2, mAccellX);
-        DeviceOrientationProfile.setY(a2, mAccellY);
-        DeviceOrientationProfile.setZ(a2, mAccellZ);
-
-        Bundle r = new Bundle();
-        DeviceOrientationProfile.setAlpha(r, mGyroX);
-        DeviceOrientationProfile.setBeta(r, mGyroY);
-        DeviceOrientationProfile.setGamma(r, mGyroZ);
-
-        DeviceOrientationProfile.setAcceleration(orientation, a1);
-        DeviceOrientationProfile.setAccelerationIncludingGravity(orientation, a2);
-        DeviceOrientationProfile.setRotationRate(orientation, r);
-        DeviceOrientationProfile.setInterval(orientation, interval);
-        return orientation;
-    }
-
-    @Override
-    public void onSensorChanged(final SensorEvent sensorEvent) {
-        if (sensorEvent.sensor.getType() == Sensor.TYPE_ACCELEROMETER) {
-
-            mAccellX = sensorEvent.values[0];
-            mAccellY = sensorEvent.values[1];
-            mAccellZ = sensorEvent.values[2];
-
-            Bundle orientation = createOrientation();
-
-            List<Event> events = EventManager.INSTANCE.getEventList(mServiceId,
-                    DeviceOrientationProfile.PROFILE_NAME, null,
-                    DeviceOrientationProfile.ATTRIBUTE_ON_DEVICE_ORIENTATION);
-
-            for (int i = 0; i < events.size(); i++) {
-                Event event = events.get(i);
-                Intent intent = EventManager.createEventMessage(event);
-                intent.putExtra(DeviceOrientationProfile.PARAM_ORIENTATION,
-                        orientation);
-                getContext().sendBroadcast(intent);
-            }
-
-            mAccelStartTime = System.currentTimeMillis();
-        } else if (sensorEvent.sensor.getType() == Sensor.TYPE_GYROSCOPE) {
-            mGyroX = sensorEvent.values[0];
-            mGyroY = sensorEvent.values[1];
-            mGyroZ = sensorEvent.values[2];
-        }
-    }
 
     //
     // File Descriptor Profile
@@ -1620,10 +1405,6 @@ public class HostDeviceService extends DConnectMessageService implements
             }
         }).start();
 
-    }
-
-    @Override
-    public void onAccuracyChanged(final Sensor sensor, final int accuracy) {
     }
 
     @Override
